# api.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /transactions:
    get:
      description: Finds transactions that match the filter.
      tags:
        - transaction-controller
      operationId: findFiltered
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: clientId to filter for.
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: year for date filter, along with month.
        - name: month
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: month for date filter.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 100
          description: limit for return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          description: offset for paging.
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionDto"
            text/csv:
              schema:
                type: string
              example: |
                id,clientId,merchantId,date,amount
        "204":
          description: No Content - nothing found in the database
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid month"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
    post:
      description: Creates a new transaction.
      tags:
        - transaction-controller
      operationId: create
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionWriteDto"
        description: Transaction (post dto) to create
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDto"
            description: Saved transaction
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
    delete:
      description: Deletes transactions that match the filter.
      tags:
        - transaction-controller
      operationId: deleteFiltered
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: clientId to filter for.
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: year for date to filter, along with month.
        - name: month
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: month to filter.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 100
          description: limit to the number of returns.
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionDto"
            text/csv:
              schema:
                type: string
              example: |
                id,clientId,merchantId,date,amount
        "204":
          description: No Content - nothing found in the database
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
  /transactions/{id}:
    get:
      description: Gets the transaction with the ID.
      tags:
        - transaction-controller
      operationId: getById
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: ID of the transaction to get.
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDto"
            text/csv:
              schema:
                type: string
              example: |
                id,clientId,merchantId,date,amount
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "ID is a required field"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "404":
          description: Not Found - resource not found
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Transaction not found with id = (id)"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
    put:
      description: Updates the existing transaction.
      tags:
        - transaction-controller
      operationId: update
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: ID to update
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionWriteDto"
            description: Transaction details to write.
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDto"
              description: The updated transaction
            text/csv:
              schema:
                type: string
              example: |
                id,clientId,merchantId,date,amount
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Missing data"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "404":
          description: Not Found – transaction does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path

    delete:
      description: Deletes the transaction with the id.
      tags:
        - transaction-controller
      operationId: deleteById
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: ID of the transaction to delete
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDto"
              description: Deleted transaction.
            text/csv:
              schema:
                type: string
              example: |
                id,clientId,merchantId,date,amount
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "ID is a required parameter"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "404":
          description: Not Found - resource not found
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Transaction not found with id (id)"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
  /transactions/total:
    description: Gets the total amount for transactions matching filters.
    get:
      tags:
        - transaction-controller
      operationId: getTotalAmount
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State to filter for.
        - name: month
          in: query
          required: true
          schema:
            type: integer
            format: int32
            maximum: 12
            minimum: 1
          description: Month to filter for.
        - name: batchSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: How many transaction entries to count.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          description: Together with batchSize for paging.
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TotalAmountDto"
        "204":
          description: No Content
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "State, month, and batch size are required parameters"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
  /transactions/tb:
    description: Returns the top or bottom of the list
    get:
      tags:
        - transaction-controller
      operationId: getMaxAmount
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
          description: Content return type header.
        - name: startYear
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: Start year to search from (inclusive).
        - name: endYear
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: End year to search to (inclusive).
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            format: int32
          description: Number of entries to get.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          description: Number of elements to skip.
        - name: dir
          in: query
          required: true
          schema:
            type: string
          description: TOP or BOTTOM to decide the order of the transactions.
        - name: acceptParam
          in: query
          required: false
          schema:
            type: string
          description: Content return type parameter.
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaximumAmountDto"
            text/csv:
              schema:
                type: string
              example: |
                maxAmount,timestamp,location
        "204":
          description: No Content
        "400":
          description: Bad request - required field missing
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Please provide a valid ordering param [top | bottom"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
        "500":
          description: Internal Server Error - unexpected database or server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                message: "Unexpected server error"
            text/csv:
              schema:
                type: string
              example: |
                timestamp,status,error,message,path
...
components:
  schemas:
...
    MaximumAmountDto:
      type: object
      properties:
        maxAmount:
          type: number
          format: double
          description: The maximum USD amount of a transaction.
        timestamp:
          type: string
          format: date-time
          description: The timestamp for the transaction.
        location:
          type: string
          description: The city of the retailer, or "ONLINE".
      description: A DTO for the maximum amount of a transaction.
...
    TotalAmountDto:
      type: object
      required:
        - day
      properties:
        day:
          type: string
          format: date-time
          description: The day of the transaction.
        totalAmount:
          type: number
          format: double
          description: The total amount for this day
      description: A DTO for the total amount in USD in transactions for a given day.
    TransactionDto:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
          description: Unique id of the transaction.
        merchantId:
          type: integer
          format: int32
          description: The ID of the merchant this transaction belongs to.
        date:
          type: string
          format: date-time
          description: The date and time when this transaction occurred.
        amount:
          type: number
          format: double
          description: The value of the transaction in USD.
        clientId:
          type: integer
          format: int32
          description: The ID of the client this transaction belongs to.
      description: A DTO for all transaction data.
    TransactionWriteDto:
      type: object
      properties:
        merchantId:
          type: number
          format: int32
          description: The ID of the merchant this transaction belongs to.
        date:
          type: string
          format: date-time
          description: The date and time when this transaction occurred.
        amount:
          type: number
          format: double
          description: The value of the transaction in USD.
        clientId:
          type: integer
          format: int32
          description: The ID of the client this transaction belongs to.
      description: A DTO for writing new transaction data.
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string